backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
gunicorn==21.2.0
mongoengine==0.29.1
pymongo==4.15.5
yfinance==0.2.66
pandas==2.3.3
python-dotenv==1.2.1 